Update EmotionsCheckin.js
I moved the `handleClick` function to the image instead of the button. 

This fixes an issue where if you click the edge of the button, the current emotion updates to a blank image because `handleclick` is looking for the `src` of the button (and not the image inside the button). 

Moving `handleClick` to the image only makes it so that nothing happens when clicking the very edge of the button. This may not be the best UX (the user is clicking the edge of the button and nothing happens) - so this could be improved in the future.